chore: update channel counter type to u64

+++ contracts/cosmwasm-vm/cw-ibc-core/tests/test_channel.rs
@@ -64,7 +64,7 @@ fn test_add_channel() {
 fn test_channel_sequence_initialisation() {
     let ctx = CwIbcCoreContext::new();
     let mut mock_deps = deps();
-    let _store = ctx.init_channel_counter(mock_deps.as_mut().storage, u128::default());
+    let _store = ctx.init_channel_counter(mock_deps.as_mut().storage, u64::default());
     let result = ctx.channel_counter(mock_deps.as_ref().storage);
 
     assert_eq!(0, result.unwrap());
